+++ web/wp-content/plugins/oshine-modules/includes/modules/newsletter.php
@@ -41,7 +41,7 @@ if (!function_exists('oshine_newsletter')) {
 	    	$output .= '<div '.$css_id.' class="oshine-mc-wrap '.$custom_class_name.' oshine-module align-'.$alignment.' '.$animate.' '.$visibility_classes.' '.$css_classes.' clearfix" '.$data_animations.' data-consent-error = "'.__('Please check the consent box, in order for us to process this form', 'oshine-modules').'">';
 	    	$output .= '<form method="POST" class="oshine-mc-form">';
 	    	$output .= '<div class="clearfix">';
-	    	$output .= '<input type="hidden" name="api_key" value="'.$api_key.'" /><input type="hidden" name="list_id" value="'.$id.'" />';
+	    	$output .= '<input type="hidden" name="api_key" value="'.$api_key.'" /><input type="hidden" name="list_id" value="'.$id.'" /><input type="hidden" name="oshine_nonce" value="' .wp_create_nonce( 'oshine_newsletter' ). '" />';
 			$output .= '<fieldset class="contact_fieldset oshine-mc-field" style="width: '.$width.'%;"><input type="text" name="email" placeholder="'.__('Email','oshine-module').'" /><div class="clear"></div></fieldset>';
 			if( !empty( $be_themes_data['consent_checkboxes'] ) ) {
 				$output .= '<fieldset class="field_consent contact_consent">	<input type="checkbox" name="contact_consent" class="consent-checkbox" placeholder="'.__('Subject','oshine-modules').'" '.$border_width.' /><span class="consent-message">'.sprintf( __('By checking this box, you consent and confirm your subscription to our newsletter. For more info check our <a href="%s" target="_blank">privacy policy</a> where you will get more info on where, how and why we store your data.', 'oshine-modules'), esc_url( $privacy_policy_link) ).'</span></fieldset>';