+++ web/wp-content/plugins/revslider/admin/includes/shortcode_generator/elementor/elementor-widget.class.php
@@ -141,13 +141,15 @@ class RevSliderElementorWidget extends \Elementor\Widget_Shortcode {
 		
 		if ( \Elementor\Plugin::$instance->editor->is_edit_mode() ) $SR_GLOBALS['loaded_by_editor'] = true;
 
+		$f = RevSliderGlobals::instance()->get('RevSliderFunctions');
+		
 		$shortcode = $this->get_settings_for_display( 'shortcode' );
 		$wrapperid = $this->get_settings_for_display( 'wrapperid' );
-		$wrapperid = empty($wrapperid) ? '': 'id="' . $wrapperid . '" ';
+		$wrapperid = empty($wrapperid) ? '': 'id="' . $f->filter_class_name($wrapperid) . '" ';
 		$shortcode = do_shortcode( shortcode_unautop( $shortcode ) );
 
 		$zindex = $this->get_settings_for_display( 'zindex' );
-		$style = $zindex ? ' style="z-index:'.$zindex.';"' : '';
+		$style = $zindex ? ' style="z-index:'.esc_attr($zindex).';"' : '';
 
 		// hack to make sure object library only opens when the user manually adds a slider to the page
 		if(empty($shortcode)) {